requirements.txt
fastapi
uvicorn
sqlalchemy
passlib[bcrypt]
python-jose[cryptography]
python-dotenv
python-multipart
psycopg2-binary

# PDFs
PyMuPDF
reportlab

# LangChain + Google GenAI (usados em main.py e api.py)
langchain
langchain-core
langchain-community
langchain-google-genai
